refactor: Move options to new file
